Final Tesco GenAI Creative Compliance Studio/requirements.txt
streamlit==1.28.0
Pillow==10.0.1
pytest==7.4.0
numpy==1.24.0
pandas==2.0.0
openpyxl==3.1.2
python-multipart==0.0.6
altair==5.1.0
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
streamlit-lottie==0.0.4
streamlit-option-menu==0.3.6
streamlit-tags==1.2.7
streamlit-image-comparison==0.0.4
opencv-python==4.8.1.78
scikit-image==0.21.0
scikit-learn==1.3.0
pytesseract==0.3.10
streamlit-elements==0.1.0